Addinng non-dms values.
Value now has all optional degrees and minutes plus also a integer
value.

The idea is that some values can't be dms, eg. to divide.

9d3m0s / 3 = 3d1m0s, but you cannot divide dms by dms.

So when starting an Divide expression, subsequent entries only
allow integers.

This adds two tests, a wrong that tests the current incorrect
expression building. And a right that is how it should be.